Add write protection for existing classes
This PR adds detection if a class already exists. If it does it will
throw an exception and won't overwrite the existing middleware.

+++ src/CreateMiddleware/CreateMiddleware.php
@@ -58,6 +58,10 @@ EOS;
             self::CLASS_SKELETON
         );
 
+        if (is_file($path)) {
+            throw CreateMiddlewareException::classExists($path, $class);
+        }
+
         file_put_contents($path, $content);
         return $path;
     }
